+++ api-clientes/database/migrations/2025_01_07_160050_create_respostas_clientes_table.php
@@ -11,12 +11,15 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('respostas_clientes', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('cliente_id')->constrained('clientes')->onDelete('cascade');
-            $table->foreignId('resposta_id')->constrained('respostas')->onDelete('cascade');
-            $table->timestamps();
-        });
+        if(!Schema::hasTable('respostas_clientes')){
+            Schema::create('respostas_clientes', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('cliente_id')->constrained('clientes')->onDelete('cascade');
+                $table->foreignId('resposta_id')->constrained('respostas')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
+        
     }
 
     /**
